Change responsibility of Board and ChessCalculator
이전에 ChessCalculator를 Board 안에서 사용했는데, 이제 굳이 그럴 필요가 없으므로 밖에서 별개로 조합하여 사용할 수 있도록 수정. 수정을 위해 groupingByCalculablePiece를 Board로 옮김. 수정 후 getScoreOf 삭제
