--- src/main/java/kr/codesquad/freddie/chess/utils/ChessCalculatorImpl.java
@@ -1,33 +1,0 @@
-package kr.codesquad.freddie.chess.utils;
-
-import kr.codesquad.freddie.chess.board.Board;
-import kr.codesquad.freddie.chess.board.File;
-import kr.codesquad.freddie.chess.piece.CalculablePiece;
-import kr.codesquad.freddie.chess.piece.Color;
-import kr.codesquad.freddie.chess.piece.Kind;
-
-import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-public class ChessCalculatorImpl implements ChessCalculator {
-    @Override
-    public double calculate(List<File> files, Color color) {
-        return groupingByCalculablePiece(files, color).entrySet().stream()
-                .mapToDouble(calculablePieceDoubleEntry -> {
-                    double point = calculablePieceDoubleEntry.getValue();
-                    Kind kind = calculablePieceDoubleEntry.getKey().getKind();
-                    return kind == Kind.PAWN && 1 < point ? point / 2 : point;
-                }).sum();
-    }
-
-    public Map<CalculablePiece, Double> groupingByCalculablePiece(List<File> files, Color color) {
-        return files.stream()
-                .flatMap(File::getCalculablePieces)
-                .filter(calculablePiece -> calculablePiece.getColor() == color)
-                .collect(Collectors.groupingBy(Function.identity(),
-                        Collectors.summingDouble(CalculablePiece::getPoint)
-                ));
-    }
-}
